+++ NEWS.md
@@ -1,7 +1,7 @@
 # glebrt 0.6.0
 
 - Add tests for logical vectors: `some()` and `same()`
-- Set up testing
+- Check function arguments and set up testing
 
 # glebrt 0.5.0